% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WashUthesis.R
\name{WashUthesis}
\alias{WashUthesis}
\title{WashU Thesis R Markdown Output Format}
\usage{
WashUthesis(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[bookdown]{pdf_document2}}}
}
\value{
An R Markdown output format
}
\description{
WashU Thesis R Markdown Output Format
}
